Create Dockerfile
Create Dockerfile based on Debian image with postfwd installed.

Postfwd is run as master process which spawns childs and
then exits. This is not designed as docker friendly. Because
of that this commit uses entrypoint script, which can override
default CMD arguments and in the end executes sleep infinity
as placeholder for keeping container alive.

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>